+++ test/analyticsTest.php
@@ -22,5 +22,6 @@ class AnalyticsTest extends PHPUnit_Framework_TestCase {
 
     $this->assertTrue($identified);
   }
+
 }
 ?>