+++ HISTORY.md
@@ -1,3 +1,9 @@
+0.1.1 / 1-30-2013
+=============
+* Adding fork consumer
+* Adding fork_queue consumer
+* Setting fork_queue consumer to be the default.
+
 0.1.0 / 1-29-2013
 =============